db: rename enterprise/internal/db package to enterprise/internal/database
