+++ enterprise/internal/database/db_test.go
@@ -1,9 +1,10 @@
-package db
+package database
 
 import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+
 	"github.com/sourcegraph/sourcegraph/internal/database/dbtesting"
 )
