+++ enterprise/internal/database/external_services_test.go
@@ -1,4 +1,4 @@
-package db
+package database
 
 import (
 	"context"
