+++ enterprise/internal/database/integration_test.go
@@ -1,4 +1,4 @@
-package db
+package database
 
 import (
 	"flag"
